config: Add config generating when no config exists.

+++ src/libgpu/gpu_config.h
@@ -11,8 +11,7 @@ namespace config
 extern bool debug;
 
 //! OpenGL debug message IDs to filter out
-// TODO: should really be a std::set, but cereal doesn't support those...
-extern std::vector<unsigned> debug_filters;
+extern std::vector<int64_t> debug_filters;
 
 //! Dump shaders
 extern bool dump_shaders;
